add event producer

+++ reservation-service/Controllers/BookingController.cs
@@ -1,4 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
+using reservation_service.Events;
 using reservation_service.Models;
 using reservation_service.Models.DTO;
 using reservation_service.Services;
@@ -10,10 +11,12 @@ namespace reservation_service.Controllers;
 public class BookingController : ControllerBase
 {
     private readonly IBookingService _bookingService;
+    private readonly IEventProducer _eventProducer;
 
-    public BookingController(IBookingService bookingService)
+    public BookingController(IBookingService bookingService, IEventProducer eventProducer)
     {
         _bookingService = bookingService;
+        _eventProducer = eventProducer;
     }
 
     [HttpGet]
@@ -39,6 +42,14 @@ public class BookingController : ControllerBase
         try
         {
             _bookingService.Create(BookingDTOMapper.RequestToBooking(id, booking));
+            _eventProducer.Publish("BookingCreated", new BookingCreated
+            {
+                Id = id,
+                OfferId = booking.OfferId,
+                FirstClassSeats = booking.FirstClassSeats,
+                SecondClassSeats = booking.SecondClassSeats,
+                Price = booking.Price
+            });
         } 
         catch(ArgumentException e)
         {
@@ -50,7 +61,19 @@ public class BookingController : ControllerBase
     [HttpDelete("{id}")]
     public IActionResult Delete(string id)
     {
+        Booking? booking = _bookingService.GetBooking(id);
+        if (booking == null)
+        {
+            return NotFound();
+        }
         _bookingService.Delete(id);
+        _eventProducer.Publish("BookingCancelled", new BookingCancelled
+        {
+            Id = id,
+            OfferId = booking.OfferId,
+            FirstClassSeats = booking.FirstClassSeats,
+            SecondClassSeats = booking.SecondClassSeats
+        });
         return NoContent();
     }
 }
